refactor: 컴포넌트 이름 및 파일 구조 변경
